Refactor interfaces out to standard location
